tools/tools.py
from langchain_core.tools import tool
from bs4 import BeautifulSoup
# 推荐使用 playwright 来处理动态加载的网站
from playwright.sync_api import sync_playwright


@tool
def scrape_xiachufang_recipe(url: str) -> str:
    """
    爬取指定的URL并提取主要的食谱内容。
    只返回清洗后的文本内容，用于LLM分析。
    """
    try:
        with sync_playwright() as p:
            browser = p.chromium.launch(headless=False)
            page = browser.new_page()
            url = "http://www.xiachufang.com/recipe/12345/"
            print("正在访问:", url)
            page.goto(url, timeout=30000)  # 30秒超时
            # 这里可以等待特定选择器出现，确保页面加载完成
            # page.wait_for_selector('.recipe-content-class') # 替换为实际的class
            html_content = page.content()
            browser.close()

        # 使用BeautifulSoup清洗HTML
        soup = BeautifulSoup(html_content, 'html.parser')
        # ... 这里是你的核心清洗逻辑 ...
        # 比如找到食谱标题、用料、步骤的<div>，提取它们的文本
        # 不同的网站需要不同的解析规则
        # target_div = soup.find('div', class_='recipe-body')
        # return target_div.get_text(separator='\n', strip=True) if target_div else "未找到食谱内容"
        # 伪代码：
        return f"成功爬取 {url} 的内容，包含食材、步骤等..."
    except Exception as e:
        return f"爬取失败: {e}"
