requirements.txt
langgraph~=0.6.4
langchain
beautifulsoup4~=4.13.4
playwright~=1.54.0
python-dotenv~=1.1.1
lxml
dashscope
langchain-community~=0.3.27
langchain-core~=0.3.74
langchain-openai~=0.3.29